requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
configparser==5.0.2
dacite==1.6.0
dataclasses-json==0.5.4
Flask==2.0.1
httpretty==1.1.3
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.1
python-dotenv==0.17.1
requests==2.25.1
requests-oauthlib==1.3.0
soupsieve==2.2.1
stringcase==1.2.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
Werkzeug==2.0.1
